backend/backend/requirements.txt
alembic==1.10.4
beautifulsoup4==4.12.2
celery==5.2.7
colorama==0.4.6
cryptography==41.0.4
dataclasses-json==0.5.7
email-validator==2.0.0.post2
emails==0.6
fastapi==0.103.2
fastapi-cache==0.1.0
grequests==0.6.0
httptools==0.5.0
langdetect==1.0.9
numexpr==2.8.4
openai==0.27.6
openapi-schema-pydantic==1.2.4
osintbuddy==0.0.4rc46.post1
passlib==1.7.4
pathspec==0.11.1
pip==23.3
pipdeptree==2.13.0
psycopg2-binary==2.9.6
pydantic-settings==2.0.3
pyfiglet==0.8.post1
PySocks==1.7.1
pytest==7.3.1
python-jose==3.3.0
python-multipart==0.0.6
python3-nmap==1.6.0
redis==4.5.4
sentry-sdk==1.32.0
sqlalchemy-json==0.7.0
tenacity==8.2.2
termcolor==2.3.0
ujson==5.8.0
undetected-chromedriver==3.4.7
uvicorn==0.22.0
validators==0.20.0
watchfiles==0.19.0
gremlinpy @ git+https://github.com/jerlendds/gremlinpy.git@7d3033e6a55ed9cb1f982ec3b58ca233e01c58e3
